removed surrogate method from Series class
