+++ pyleoclim/tests/test_core_Series.py
@@ -537,66 +537,7 @@ class TestSel:
             ts.sel(time=1, value=1)
 
 
-class TestUISeriesSurrogates:
-    ''' Test Series.surrogates()
-    '''
-    def test_surrogates_t0(self, p=2, eps=0.2):
-        ''' Generate AR(1) surrogates based on a AR(1) series with certain parameters,
-        and then evaluate and assert the parameters of the surrogates are correct
-        '''
-        g = 0.5  # persistence
-        ar = [1, -g]
-        ma = [1, 0]
-        n = 1000
-        ar1 = arma_generate_sample(ar, ma, nsample=n, scale=1)
-        ts = pyleo.Series(time=np.arange(1000), value=ar1)
 
-        ts_surrs = ts.surrogates(number=p)
-        for ts_surr in ts_surrs.series_list:
-            g_surr = tsmodel.ar1_fit(ts_surr.value)
-            assert np.abs(g_surr-g) < eps
-            
-    @pytest.mark.parametrize('number',[1,5])     
-    def test_surrogates_uar1_match(self, number):
-        ts = gen_ts(nt=550, alpha=1.0)
-        # generate surrogates
-        surr = ts.surrogates(method = 'uar1', number = number, time_pattern ="match")
-        for i in range(number):
-            assert(np.allclose(surr.series_list[i].time, ts.time))
-            
-    def test_surrogates_uar1_even(self, p=5):
-        ts = gen_ts(nt=550, alpha=1.0)
-        time_incr = np.median(np.diff(ts.time))
-        # generate surrogates
-        surr = ts.surrogates(method = 'uar1', number = p, time_pattern ="even", settings={"time_increment" :time_incr})
-        for i in range(p):
-            assert(np.allclose(tsmodel.inverse_cumsum(surr.series_list[i].time),time_incr))
-
-    def test_surrogates_uar1_random(self, p=5, tol = 0.5):
-        tau = 2
-        sigma_2 = 1
-        n = 500
-        # generate time index
-        t = np.arange(1,(n+1))
-        # create time series
-        ys = tsmodel.uar1_sim(t, tau_0=tau, sigma_2_0=sigma_2)
-        ts = pyleo.Series(time = t, value=ys, auto_time_params=True,verbose=False)
-        # generate surrogates default is exponential with parameter value 1
-        surr = ts.surrogates(method = 'uar1', number = p, time_pattern ="random")
-        #surr = ts.surrogates(method = 'uar1', number = p, time_pattern ="uneven",settings={"delta_t_dist" :"poisson","param":[1]} )
-    
-        for i in range(p):
-            delta_t = tsmodel.inverse_cumsum(surr.series_list[i].time)
-            # Compute the empirical cumulative distribution function (CDF) of the generated data
-            empirical_cdf, bins = np.histogram(delta_t, bins=100, density=True)
-            empirical_cdf = np.cumsum(empirical_cdf) * np.diff(bins)
-            # Compute the theoretical CDF of the Exponential distribution
-            theoretical_cdf = expon.cdf(bins[1:], scale=1)
-            # Trim theoretical_cdf to match the size of empirical_cdf
-            theoretical_cdf = theoretical_cdf[:len(empirical_cdf)]
-            # Compute the L2 norm (Euclidean distance) between empirical and theoretical CDFs
-            l2_norm = np.linalg.norm(empirical_cdf - theoretical_cdf)
-            assert(l2_norm<tol)
             
 
 class TestUISeriesSummaryPlot:
